move selector button to separate file
