+++ src/components/swap/components/trade-input-selector.tsx
@@ -1,12 +1,10 @@
-import Image from 'next/image'
-
-import { ChevronDownIcon } from '@chakra-ui/icons'
 import { Flex, Input, Text } from '@chakra-ui/react'
 
 import { Token } from '@/constants/tokens'
 import { colors } from '@/lib/styles/colors'
 
 import { Caption } from './caption'
+import { SelectorButton } from './selector-button'
 
 interface TradeInputSelectorConfig {
   isInputDisabled?: boolean
@@ -162,29 +160,5 @@ const PriceUsd = (props: PriceUsdProps) => (
         &nbsp;{props.priceImpact.value}
       </Text>
     )}
-  </Flex>
-)
-
-type SelectorProps = {
-  onClick: () => void
-  image: string
-  symbol: string
-}
-
-const SelectorButton = ({ image, symbol, onClick }: SelectorProps) => (
-  <Flex
-    align='center'
-    bg={colors.icGray1}
-    borderRadius='32'
-    cursor='pointer'
-    onClick={onClick}
-    p='10px'
-    shrink={0}
-  >
-    <Image alt={`${symbol} logo`} src={image} width={20} height={20} />
-    <Text color={colors.icBlack} fontSize={'14px'} fontWeight={500} mx='8px'>
-      {symbol}
-    </Text>
-    <ChevronDownIcon w={6} h={6} color={colors.icGray4} />
   </Flex>
 )
