+++ src/components/swap/components/trade-output.tsx
@@ -1,6 +1,3 @@
-import Image from 'next/image'
-
-import { ChevronDownIcon } from '@chakra-ui/icons'
 import { Flex, Text } from '@chakra-ui/react'
 
 import { Token } from '@/constants/tokens'
@@ -10,6 +7,7 @@ import { colors } from '@/lib/styles/colors'
 import { FormattedQuoteDisplay } from '../hooks/use-swap/formatters/result'
 import { Caption } from './caption'
 import { QuoteResult } from './quote-result'
+import { SelectorButton } from './selector-button'
 
 interface TradeOutputProps {
   caption: string
@@ -58,27 +56,3 @@ export const TradeOutput = (props: TradeOutputProps) => {
     </Flex>
   )
 }
-
-type SelectorProps = {
-  onClick: () => void
-  image: string
-  symbol: string
-}
-
-const SelectorButton = ({ image, symbol, onClick }: SelectorProps) => (
-  <Flex
-    align='center'
-    bg={colors.icGray1}
-    borderRadius='32'
-    cursor='pointer'
-    onClick={onClick}
-    p='10px'
-    shrink={0}
-  >
-    <Image alt={`${symbol} logo`} src={image} width={20} height={20} />
-    <Text color={colors.icBlack} fontSize={'14px'} fontWeight={500} mx='8px'>
-      {symbol}
-    </Text>
-    <ChevronDownIcon w={6} h={6} color={colors.icGray4} />
-  </Flex>
-)
